refactor(#19): make flex style function
