+++ src/components/GlobalLayout/index.tsx
@@ -1,12 +1,9 @@
+import { flex } from '@/styles/common/flex';
 import { create, props } from '@stylexjs/stylex';
 import type { PropsWithChildren } from 'react';
 
 const globalLayoutStyles = create({
   main: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-
     minHeight: '100vh',
 
     background: 'hsl(220deg 14.3% 95.9%)',
@@ -14,5 +11,5 @@ const globalLayoutStyles = create({
 });
 
 export const GlobalLayout = ({ children }: PropsWithChildren) => {
-  return <div {...props(globalLayoutStyles.main)}>{children}</div>;
+  return <div {...props(globalLayoutStyles.main, flex.center)}>{children}</div>;
 };
