+++ src/pages/login/components/Card/Content.tsx
@@ -1,3 +1,4 @@
+import { flex } from '@/styles/common/flex';
 import { create, props } from '@stylexjs/stylex';
 import type { PropsWithChildren } from 'react';
 
@@ -5,14 +6,11 @@ const contentStyles = create({
   content: {
     width: '20rem',
 
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
     flexDirection: 'column',
     rowGap: '1rem',
   },
 });
 
 export const Content = ({ children }: PropsWithChildren) => {
-  return <div {...props(contentStyles.content)}>{children}</div>;
+  return <div {...props(contentStyles.content, flex.center)}>{children}</div>;
 };
