+++ src/pages/login/components/Card/Footer.tsx
@@ -1,6 +1,6 @@
 import { create, props } from '@stylexjs/stylex';
 import type { PropsWithChildren } from 'react';
-import { weight } from '../../../../assets/tokens/font.stylex';
+import { weight } from '../../../../styles/tokens/font.stylex';
 
 const footerStyles = create({
   footer: {
