+++ src/styles/common/flex.ts
@@ -1,0 +1,9 @@
+import { create } from '@stylexjs/stylex';
+
+export const flex = create({
+  center: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
